integrate availability selector
